requirements.txt
Flask
line-bot-sdk
python-decouple
pymongo==2.3
opencv-python-headless==4.2.0.32
pytesseract==0.3.9
pillow
dnspython